Move out of Indexer namespace
