+++ administrator/components/com_finder/Helper/ResponseHelper.php
@@ -7,7 +7,7 @@
  * @license     GNU General Public License version 2 or later; see LICENSE.txt
  */
 
-namespace Joomla\Component\Finder\Administrator\Indexer;
+namespace Joomla\Component\Finder\Administrator\Helper;
 
 defined('_JEXEC') or die;
 
@@ -21,7 +21,7 @@ use Joomla\CMS\Log\Log;
  *
  * @since  2.5
  */
-class Response
+class ResponseHelper
 {
 	/**
 	 * Class Constructor
